Population count per residential building in Singapore

Example (random 5% subset) dataset of residential building polygons in Singapore, each with a population count (column popcount) estimated via dasymetric mapping.

buildings_pop_sgp

Format

sf polygons.

Source

Building polygons copyrighted OpenStreetMap contributors and available from https://www.openstreetmap.org. Made available via the ODbL License.

Population data and census block polygons data(pop_sgp) are from the Department of Statistics Singapore; and Singapore Master Plan Subzones, respectively; land use polygons data(landuse_sgp) are from the Singapore Land Use Master Plan released in 2019. All are made available under the terms of the Singapore Open Data Licence version 1.0.

Details

Building polygons were downloaded from OpenStreetMap (data snapshot on 2021-01-01 from the Geofabrik database), using the function get_buildings_osm(). The population count per census block in the year 2020 was re-distributed across the buildings located within residential land use zones, by performing dasymetric mapping using the functions pop_dasymap() and pop_density_polygonise(). See vignette and examples in pop_density_polygonise() for more details. The dataset is a random 5% subset of the resulting polygons.

Examples

data(buildings_pop_sgp)
head(buildings_pop_sgp)

#> Simple feature collection with 6 features and 1 field
#> Geometry type: POLYGON
#> Dimension:     XY
#> Bounding box:  xmin: 103.8412 ymin: 1.297955 xmax: 103.8968 ymax: 1.426561
#> Geodetic CRS:  WGS 84
#>    popcount                       geometry
#> 1 32.367111 POLYGON ((103.8412 1.420676...
#> 2 20.880897 POLYGON ((103.8808 1.320019...
#> 3  2.311594 POLYGON ((103.8643 1.320283...
#> 4 56.747406 POLYGON ((103.8504 1.303723...
#> 5  6.368404 POLYGON ((103.8516 1.426561...
#> 6  9.502521 POLYGON ((103.8966 1.298226...