Population counts within census block polygons in Singapore

Example dataset containing Singapore census data for years 2018 and 2020 from the Department of Statistics Singapore, joined by name to the Master Plan Subzone polygons from data.gov.sg. Both datasets are released under the Singapore Open Data License.

pop_sgp

Format

sf polygons with census data for years 2018 and 2020. Data is in the 'long' format (rows are repeated for each census year). The following columns are used in this package:

subzone_n

Census block name

year

Census year

pop_count

Population count within census block

Source

Contains census data from the Department of Statistics Singapore and polygons from the Singapore Master Plan Subzones, both of which are made available under the terms of the Singapore Open Data Licence version 1.0.

Examples

data(pop_sgp)
head(pop_sgp)

#> Simple feature collection with 6 features and 12 fields
#> Geometry type: MULTIPOLYGON
#> Dimension:     XY
#> Bounding box:  xmin: 103.8303 ymin: 1.280459 xmax: 103.8811 ymax: 1.296311
#> Geodetic CRS:  WGS 84
#> # A tibble: 6 x 13
#>   subzone_n   year pop_count                       geometry subzone_no subzone_c
#>   <chr>      <dbl>     <dbl>             <MULTIPOLYGON [°]> <chr>      <chr>    
#> 1 Marina Ea…  2018         0 (((103.8802 1.283859, 103.880… 1          MESZ01   
#> 2 Marina Ea…  2020         0 (((103.8802 1.283859, 103.880… 1          MESZ01   
#> 3 Instituti…  2018      3470 (((103.8376 1.295599, 103.837… 5          RVSZ05   
#> 4 Instituti…  2020      3140 (((103.8376 1.295599, 103.837… 5          RVSZ05   
#> 5 Robertson…  2018      2720 (((103.8341 1.292476, 103.834… 1          SRSZ01   
#> 6 Robertson…  2020      2990 (((103.8341 1.292476, 103.834… 1          SRSZ01   
#> # … with 7 more variables: ca_ind <chr>, pln_area_n <chr>, pln_area_c <chr>,
#> #   region_n <chr>, region_c <chr>, inc_crc <chr>, fmel_upd_d <chr>