Land use master plan for the city of Singapore

Example dataset containing Singapore Master Plan Land Use Zones for the years 2014 and 2019. Released under the Singapore Open Data License.

landuse_sgp

Format

sf polygons of land use released in years 2014 and 2019. Data is in the 'long' format (rows are repeated for each year). The following columns are used in this package:

lu_desc

Land use type (description)

year

Year of Master Plan

Source

Polygons from the Singapore Land Use Master Plans released in the years 2014 and 2019. Made available under the terms of the Singapore Open Data Licence version 1.0.

Examples

data(landuse_sgp)
head(landuse_sgp)
#> Simple feature collection with 6 features and 23 fields
#> Geometry type: MULTIPOLYGON
#> Dimension:     XY
#> Bounding box:  xmin: 103.7339 ymin: 1.338392 xmax: 103.8353 ymax: 1.376801
#> Geodetic CRS:  WGS 84
#> # A tibble: 6 × 24
#>   lu_desc    year objectid oid_1 lu_text gpr   inc_crc  x_addr y_addr shape_leng
#>   <chr>     <dbl>    <int> <int> <chr>   <chr> <chr>     <dbl>  <dbl>      <dbl>
#> 1 UTILITY    2014        1     0 U       EVA   BF0C373… 23227. 37189.      1475.
#> 2 OPEN SPA…  2014        2     0 NA      EVA   DD6E9ED… 24628. 36234.       814.
#> 3 OPEN SPA…  2014        3     0 NA      EVA   9A88CBC… 28065. 35926.       771.
#> 4 OPEN SPA…  2014        4     0 NA      EVA   2481817… 26177. 35948.      1687.
#> 5 ROAD       2014        5     0 NA      NA    C56C3E0… 28179. 35988.       192.
#> 6 PLACE OF…  2014        6     0 W       EVA   CB07DF0… 16995. 39803.       359.
#> # … with 14 more variables: shape_area <dbl>, geometry <MULTIPOLYGON [°]>,
#> #   name <chr>, descriptio <chr>, timestamp <chr>, begin <chr>, end <chr>,
#> #   altitudemo <chr>, tessellate <dbl>, extrude <dbl>, visibility <dbl>,
#> #   draworder <dbl>, icon <chr>, snippet <chr>